ok-001

+++ join_us_server/mangerRouter/consult/index.js
@@ -2,6 +2,7 @@ const consultRouter = require("express").Router()
 const query = require('../../utils/mysql')
 const { returnErr } = require('../../utils/returnErr')
 const { v4: uuidv4 } = require('uuid');
+const {noAllowConsultArr } = require('../../config/noAllowConsult');
 const categoryData = [
   { id: 0, categoryName: "求职必读" },
   { id: 1, categoryName: "数据报告" },
@@ -34,19 +35,26 @@ consultRouter.post('/consult', (req, res) => {
 
 //查询资讯
 consultRouter.get('/consults', (req, res) => {
-  let { category, pageOn, pageSize } = req.query
+  let { category, pageOn, pageSize ,keyword} = req.query
   if (!pageOn || !pageSize) return returnErr(res, '参数错误')
   if (pageSize >= 10) {
     pageSize = 10
   }
+  let str = ''
   let sql = ''
   let sqlCount = ''
   if (category) {
-    sqlCount = `select count(*) from consult  inner join manger on consult.manger_id = manger.manger_id where category = '${category}' `
-    sql = `select * from consult inner join manger on consult.manger_id = manger.manger_id  where category = '${category}'  order by consult.id desc limit ${(pageOn - 1) * pageSize} ,${pageSize}`
+    if(keyword){
+      str = ` and (consult.category like '%${keyword}%' or consult.title like '%${keyword}%')  `
+    }
+    sqlCount = `select count(*) from consult  inner join manger on consult.manger_id = manger.manger_id where category = '${category}' ${str} `
+    sql = `select * from consult inner join manger on consult.manger_id = manger.manger_id  where category = '${category}' ${str} order by consult.id desc limit ${(pageOn - 1) * pageSize} ,${pageSize}`
   } else {
-    sqlCount = `select count(*) from consult inner join manger on consult.manger_id = manger.manger_id`
-    sql = `select * from consult inner join manger on consult.manger_id = manger.manger_id order by consult.id desc  limit ${(pageOn - 1) * pageSize} ,${pageSize} `
+    if(keyword){
+      str = `where consult.category like '%${keyword}%' or consult.title like '%${keyword}%'`
+    }
+    sqlCount = `select count(*) from consult inner join manger on consult.manger_id = manger.manger_id ${str}`
+    sql = `select * from consult inner join manger on consult.manger_id = manger.manger_id  ${str} order by consult.id desc  limit ${(pageOn - 1) * pageSize} ,${pageSize} `
   }
   query(sqlCount, countResult => {
     query(sql, result => {
@@ -76,6 +84,8 @@ consultRouter.post('/consult/state', (req, res) => {
 })
 
 
+
+
 //修改资讯
 consultRouter.patch('/consult', (req, res) => {
   let {consult_id, category, content, cover_img, title, descs } = req.body
@@ -83,6 +93,7 @@ consultRouter.patch('/consult', (req, res) => {
   if (!consult_id || !category || !content || !title || !descs) return returnErr(res, '参数错误')
   content = content?.replaceAll("'", '"')
 
+ 
   const sql = `update consult set category = '${category}',content = '${content}',
   title = '${title}',descs = '${descs}',cover_img = '${cover_img}',updateTime = '${new Date().toLocaleString().replaceAll('/','-')}'
   where consult_id = '${consult_id}'`
@@ -100,6 +111,7 @@ consultRouter.patch('/consult', (req, res) => {
 consultRouter.delete('/consult', (req, res) => {
   const {consult_id} = req.body
   if (!consult_id) return returnErr(res, '参数错误')
+  if(noAllowConsultArr.includes(consult_id)) return returnErr(res, '该资讯内容很重要，无法删除')
   const sql = `delete from consult where consult_id = '${consult_id}'`
   query(sql, result => {
     if (result.affectedRows >= 1) {
